backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-JWT-Extended==4.6.0
pandas==2.1.4
python-dotenv==1.0.0
Werkzeug==3.0.1
flasgger==0.9.7.1